chore: rename and re-organize upgrade-manager code
