+++ src/backend/index.js
@@ -23,7 +23,7 @@ const debug = require("debug");
 const ServerStatus = require("./status");
 const constants = require("./constants");
 const createServer = require("./server");
-const { getInstallerInfo } = require("./lib/utils");
+const { getInstallerInfo } = require("./upgrade-manager/utils");
 const createBugsnag = require("@bugsnag/js");
 const semverPrerelease = require("semver/functions/prerelease");
 const { version } = require("../../package.json");
